Day10
Problems of this section is very much easy and all of those are of a modification of same type problem which mainly focused on using some commonly used function works with list, dictionary, tuple. In my entire solutions, I haven't tried to solve problems efficiently. Rather I tried to solve differently that I can. This will help a beginner to know how the simplest problems can be solved in different ways.